Now using Double instead of Float, removed battery opt permissions, fixed entry animation, top bar, menu bar, ...

+++ app/src/main/java/io/github/stcksmsh/kap/data/AppSettingsStore.kt
@@ -36,6 +36,6 @@ fun loadAppSettings(context: Context): AppSettings {
             )!!
         ),
         quickWaterAdditionVolumes = sharedPreferences.getString("quickWaterAdditionVolumes", null)
-            ?.split(';')?.map { it.toFloat() } ?: AppSettings.defaultQuickWaterAdditionVolumes
+            ?.split(';')?.map { it.toDouble() } ?: AppSettings.defaultQuickWaterAdditionVolumes
     )
 }
