+++ app/src/main/java/io/github/stcksmsh/kap/data/UserSettingsStore.kt
@@ -7,9 +7,9 @@ fun saveUserSettings(context: Context, userSettings: UserSettings) {
     val sharedPreferences = context.getSharedPreferences("user_prefs", Context.MODE_PRIVATE)
     with(sharedPreferences.edit()) {
         putInt("age", userSettings.age)
-        putFloat("weight", userSettings.weight)
+        putDouble("weight", userSettings.weight)
         putInt("dailyPhysicalActivityDuration", userSettings.dailyPhysicalActivity)
-        putFloat("dalyWaterGoal", userSettings.dailyWaterGoal)
+        putDouble("dalyWaterGoal", userSettings.dailyWaterGoal)
         apply()
     }
 }
@@ -18,9 +18,9 @@ fun loadUserSettings(context: Context): UserSettings {
     val sharedPreferences = context.getSharedPreferences("user_prefs", Context.MODE_PRIVATE)
     return UserSettings(
         age = sharedPreferences.getInt("age", 0),
-        weight = sharedPreferences.getFloat("weight", 0f),
+        weight = sharedPreferences.getDouble("weight", 0.0),
         dailyPhysicalActivity = sharedPreferences.getInt("dailyPhysicalActivityDuration", -1),
-        dailyWaterGoal = sharedPreferences.getFloat("dalyWaterGoal", 0f)
+        dailyWaterGoal = sharedPreferences.getDouble("dalyWaterGoal", 0.0)
     )
 }
 
@@ -29,11 +29,11 @@ fun hasUserSettings(context: Context): Boolean {
     return sharedPreferences.contains("age")  // Check if age or other data exists
 }
 
-fun calculateOptimalWaterIntake(age: Int, weight: Float, dailyPhysicalActivity: Int): Float {
+fun calculateOptimalWaterIntake(age: Int, weight: Double, dailyPhysicalActivity: Int): Double {
     val ageBasedWaterIntake = when {
         age < 18 -> 200
         age < 55 -> 0
         else -> -200
     }
-    return ((weight * 0.033) * 1000 + (dailyPhysicalActivity / 60.0) * 0.35 * 1000 + ageBasedWaterIntake).toFloat()
+    return ((weight * 0.033) * 1000 + (dailyPhysicalActivity / 60.0) * 0.35 * 1000 + ageBasedWaterIntake)
 }
